+++ src/RcppExports.cpp
@@ -14,10 +14,12 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 
 RcppExport SEXP _rcpp_module_boot_stan_fit4GUTS_IT_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4GUTS_SD_mod();
+RcppExport SEXP _rcpp_module_boot_stan_fit4GUTS_hb_only_mod();
 
 static const R_CallMethodDef CallEntries[] = {
     {"_rcpp_module_boot_stan_fit4GUTS_IT_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4GUTS_IT_mod, 0},
     {"_rcpp_module_boot_stan_fit4GUTS_SD_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4GUTS_SD_mod, 0},
+    {"_rcpp_module_boot_stan_fit4GUTS_hb_only_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4GUTS_hb_only_mod, 0},
     {NULL, NULL, 0}
 };