+++ NEWS.md
@@ -1,3 +1,14 @@
+# BeeGUTS 1.3
+* Parallelization of LC/LD calculations
+* Improved handling of hb values for validation and predictions
+* Minor fixes to Short time of Effect calculations
+
+# BeeGUTS 1.2.1
+* Minor fixes
+
+# BeeGUTS 1.2.0
+* Introduce calculations for Time Reinforced Toxicity test and time of effect
+
 # BeeGUTS 1.1.3
 * Update to rstan version 2.26